Changes to accessibility (#203)
Changed EventListing page <a> element to a button since <a> link element shouldn't be used for a javascript function.

Made the Form page background 50% more transparent since it was distracting from the actual main content of the form. (Might need to make it even more transparent if needed, but we'll start with this)

Updated Info/EventListing page links to have underlines so that color blind people will be able to distinguish links better.